improve dakar

+++ Dakar/src/modules/Moto.java
@@ -2,12 +2,12 @@ package modules;
 
 import java.math.BigDecimal;
 
-public class Moto extends Veiculo{
+public class Moto extends Veiculo {
 
-    private static final BigDecimal pesoMoto = BigDecimal.valueOf(300.0);
+    private static final double pesoMoto = 300.0;
     private static final int rodas = 2;
 
-    public Moto(BigDecimal velocidade, BigDecimal aceleracao, double anguloGiro, String placa) {
+    public Moto(double velocidade, double aceleracao, double anguloGiro, String placa) {
         super(velocidade, aceleracao, anguloGiro, placa, pesoMoto, rodas);
     }
